Tidy namespace hierarchy
