+++ src/DXF/Arc.php
@@ -1,26 +1,31 @@
 <?php
-namespace DXFwriter;
+namespace DXF;
 
 /**
-* 3dface
+* Arc, angles in degrees.
 * subclass of Entity
 * 
 * Used attributes
-* points (required) default null
-*
+* center default array(0,0,0)
+* radius default 1
+* startAngle default 0.1
+* endAngle default 90
 */
-
-class DxfFace extends DxfEntity{
+class DxfArc extends DxfEntity{
 
 	/*
 	* Constructor
 	* It is recommended that sublasses calls parent::__construct($attributes)
 	* after setting default attributes 
 	*
-	* @param  Array	$attributes	array of attributes
+	* @param  	Array	$attributes	array of attributes
 	*/
 	function __construct($attributes = array()){
 		$defaults = array();
+		$defaults['center'] = array(0, 0, 0);
+		$defaults['radius'] = 1;
+		$defaults['startAngle'] = 0.1;
+		$defaults['endAngle'] = 90;
 		parent::__construct(array_merge($defaults, $attributes));
 	}
 
@@ -32,9 +37,12 @@ class DxfFace extends DxfEntity{
 	*/
 	function __toString(){
 		// TODO all are string values, maybee som should be decimal
-		return sprintf("0\n3DFACE\n%s\n%s",
+		return sprintf("0\nARC\n%s\n%s40\n%f\n50\n%f\n51\n%f",
 									$this->common(),
-									points($this->attributes['points'])
+									point($this->attributes['center']),
+									$this->attributes['radius'],
+									$this->attributes['startAngle'],
+									$this->attributes['endAngle']
 				);
 	}
 }
