+++ src/DXF/Insert.php
@@ -1,5 +1,5 @@
 <?php
-namespace DXFwriter;
+namespace DXF;
 
 /**
 * Insert Block instance.
