+++ src/DXF/LwPolyLine.php
@@ -1,5 +1,5 @@
 <?php
-namespace DXFwriter;
+namespace DXF;
 
 /**
 * LwPolyLine
