+++ src/DXF/Style.php
@@ -1,5 +1,5 @@
 <?php
-namespace DXFwriter;
+namespace DXF;
 
 /*
 * Style - Text style
